Initial check in of the software
This is my initial version of the software required to run on the
Netduino Plus 2.  I am aware that there are one or two bugs which cause
the Netduino to lock up.  If you can reproduce them then please let me
know so I can fix (or submit a pull request)