Fixed crash due to longer Lat/Long than expected.
I was advised by Steve Jacobs that his monitor crashed.  We identified
the problem to be a difference in length between his Lat/Long and mine.
I have modified the code to pad accordingly whitespace when needed.
Tested and still works fine on my monitor.  Also bumped the version to
1.0.3.